requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-jwt==1.10.0
djangorestframework-simplejwt==5.2.0
idna==3.3
psycopg2==2.9.3
PyJWT==2.4.0
pyngrok==5.1.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
twilio==7.9.3
urllib3==1.26.9
